Add more commands (#120)
* Revert "Final contribution by me (#111)"

revert changes
This reverts commit af59806cee89a5eb4fff92c3f754a78cb08838b3.

* fix packages

* fix styling issues

* add splash screen

* random music

* remove unnecessary files

* jar file music

* refactor code

* fix comments and test

* fix splashscreen

* print command done

* bypass issue with runtime error

* fix checkstyle

* fix english

* fix english

* remove unused imports

* fix conflicts bug

* exit properly

* integrate shan code

* fix Vbox height

* results command

* refactor code

* fix checkstyle

* fix checkstyle
