+++ src/main/java/com/algosenpai/app/storage/UserStorageParser.java
@@ -1,10 +1,6 @@
 package com.algosenpai.app.storage;
 
 import com.algosenpai.app.stats.ChapterStat;
-
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.util.Scanner;
 
 public class UserStorageParser {
@@ -14,13 +10,9 @@ public class UserStorageParser {
 
     /**
      * Initialises everything so that Scanner works.
-     * @throws FileNotFoundException will never be thrown as the file would already be prepared.
      */
-    public UserStorageParser() throws FileNotFoundException {
-        //String workingDir = System.getProperty("user.dir") + File.separator + "UserData.txt";
-        //FileReader reader = new FileReader(workingDir);
+    public UserStorageParser() {
         this.inputScanner = new Scanner(this.getClass().getResourceAsStream("/data/UserData.txt"));
-
     }
 
     /**
